Split Core, Content & LiveText into MVI subparts
