+++ src/components/Content/index.ts
@@ -1,28 +1,11 @@
-import xs, { Stream } from "xstream"
+import xs from "xstream"
 import isolate from "@cycle/isolate"
-import { div, VNode } from "@cycle/dom"
 
 import { Sources, Sinks } from "types"
 import Model from "model"
 import CustomText from "components/CustomText"
 import LiveText from "components/LiveText"
-
-
-function view(textStatusKO$, textStatusOK$, liveTextVDom$, customTextVDom$): Stream<VNode> {
-  return xs.combine(textStatusKO$, textStatusOK$, liveTextVDom$, customTextVDom$)
-    .map(([textStatusKO, textStatusOK, liveText, customText]) => {
-      return div('.ta-content', {
-        attrs: {tabindex: 0},
-        class: {
-          'ta-text-status--failed': textStatusKO,
-          'ta-text-status--success': textStatusOK
-        }
-      }, [
-        liveText,
-        customText
-      ])
-  })
-}
+import view from "./view"
 
 
 export default function LiveText(sources: Sources): Sinks {
