+++ src/components/Core/index.ts
@@ -1,112 +1,13 @@
 import xs from "xstream"
 import isolate from "@cycle/isolate"
-import { h1, h2, div } from "@cycle/dom"
 
-import { AppState, Sources, Sinks } from "types"
-import { INITIAL_APP_STATE } from "utils"
-
+import { Sources, Sinks } from "types"
 import Content from "components/Content"
 import Metrics from "components/Metrics"
 // import ReplayTyping from "components/replay_typing"
-
-
-// Intent: computes mutation proposals based on raw events.
-function intent(sources: Sources) {
-}
-
-
-// Model: map actions to state reducers.
-function model(actions) {
-  return xs.of(function initialReducer(prevState: AppState) {
-    if (prevState)
-      return prevState
-    else
-      return INITIAL_APP_STATE
-  })
-}
-
-// View: decorates app state and re-renders in place.
-function view(contentVDom$, metricsVDom$) {
-  return xs.combine(contentVDom$, metricsVDom$)
-    .map(([content, metrics]) => {
-      return div('.typing-app.ta', [
-        div('.ta-side', [
-          h1('.ta-title', 'typometer'),
-          div('.ta-settings', [
-            h2('Settings'),
-            // h(replay_settings.sel, replay_settings.data, replay_settings.children),
-          ])
-        ]), // .ta-side
-
-        div('.ta-main', [
-
-          div('.ta-header', [
-            metrics
-
-            // div(classnames('.ta-progress .ta-progress--done', {'.u-wip': !attributes.done}), ' Done!')
-
-          ]), // .ta-header
-
-          content
-
-        ]) // .ta-main
-
-      ]) // .ta
-    })
-}
-
-
-// Next-action-predicate aka. internal side-effect handler.
-//
-// Computes any required side-effect and pushes it down the pipe.
-// A side-effect could be:
-// - a mutation proposal
-// - an emulated user-action (resulting in a mutation proposal)
-// - an effect on some external asset with no internal consequence whatsoever
-//
-// The last use-case is more suited to sources-less drivers so should be avoided.
-// Emulated user-action are better handled outside the intent() function due to
-// stream wiring. Therefore, such actions must be pre-processed here as a
-// mutation proposal and exposed as such.
-// function nap(app_state$) {
-//   // Typing is over but stats (model decoration) are still pending: trigger
-//   // updates and re-rendering with a fake keypress.
-//   const compute_wpm$ = app_state$
-//     .filter(_ => {
-//       return Model.Singleton.get().isDone()
-//         && !Model.Singleton.get().attributes.stop
-//     })
-//     .map(app_state => (new TypingAction).process())
-//
-//   // Once everything is over, including stats computation/display, let's reset
-//   // everything after a short delay so the user may retry.
-//   const auto_reset$ = app_state$
-//     .filter(_ => {
-//       return Model.Singleton.get().isDone()
-//         && Model.Singleton.get().attributes.stop
-//     })
-//     .compose(delay(2000))
-//     .map(_ => Model.Singleton.clear())
-//
-//   // User has stopped, must reset state.
-//   const manual_reset$ = app_state$
-//     .filter(_ => {
-//       return Model.Singleton.get().hasStopped()
-//     })
-//     .map(_ => Model.Singleton.clear())
-//
-//   return xs.merge(
-//     compute_wpm$,
-//     auto_reset$,
-//     manual_reset$
-//   ).startWith(undefined)
-// }
-
-
-function nap(state$) {
-  return xs.create()
-}
-
+import model from "./model"
+import view from "./view"
+import nap from "./nap"
 
 // Main: wires everything up using circular streams.
 // Note: next-action-predicates bypass the intent() layer by design.
@@ -134,8 +35,7 @@ export default function Core(sources: Sources): Sinks {
   // const nap$ = nap(app_state$)
 
   const state$ = sources.onion.state$
-  const actions = intent(sources)
-  const parentReducer$ = model(actions)
+  const parentReducer$ = model()
   const napReducer$ = nap(state$)
 
   // Content
