+++ src/components/LiveText/intent.ts
@@ -1,0 +1,11 @@
+import { Sources } from "types"
+
+export default function intent(sources: Sources) {
+  return {
+    newChar$: sources.DOM
+      .select('document').events('keydown')
+      .filter(e => !/^(Dead|F2)/.test(e.key))
+      .filter(e => !/^(Tab|Control|Alt|Shift|Meta).*/.test(e.code))
+      .map(e => e.key)
+  }
+}
