+++ src/components/LiveText/model.ts
@@ -1,0 +1,13 @@
+import { Stream } from "xstream"
+
+import { Reducer } from "types"
+import TypingAction from "actions/typing"
+
+export default function model(actions): Stream<Reducer> {
+  return actions.newChar$
+    .map(newChar => {
+      return function newCharReducer(prevState) {
+        return TypingAction(newChar, prevState)
+      }
+    })
+}
